BDN-292. Changed total_in_app_amount with Double. (#103)

+++ bidon/src/main/java/org/bidon/sdk/segment/models/SegmentAttributes.kt
@@ -7,7 +7,7 @@ internal data class SegmentAttributes(
     val age: Int?,
     val gender: Gender?,
     val customAttributes: Map<String, Any>,
-    val inAppAmount: Int?,
+    val inAppAmount: Double?,
     val isPaying: Boolean?,
     val gameLevel: Int?,
 ) {
